requirements.txt
# Core ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# MLFlow and tracking
mlflow>=2.0.0
dvc>=3.0.0

# Data processing
scipy>=1.9.0
joblib>=1.2.0

# Google Cloud and BigQuery
google-cloud-bigquery>=3.4.0
google-cloud-storage>=2.7.0
db-dtypes>=1.0.0

# Advanced ML and Time Series
torch>=2.0.0
torch-geometric>=2.4.0
pytorch-lightning>=2.0.0
transformers>=4.30.0

# Network Analysis and Graph Theory
networkx>=3.1
grakel>=0.1.8
karateclub>=1.3.0

# Time Series and Econometrics
statsmodels>=0.14.0
arch>=5.3.0  # For GARCH models
linearmodels>=4.25  # For advanced econometrics
pykalman>=0.9.5  # For state-space models

# Granger Causality and Spillover Analysis
tsfresh>=0.20.0
# pyGranger>=0.1.0  # Package not available - implement custom

# Additional Scientific Computing
spectrum>=0.8.0  # For spectral analysis
pyEMD>=1.5.0  # For EMD decomposition
scipy>=1.10.0
numba>=0.57.0  # For performance optimization
yfinance>=0.2.0  # For benchmark data (Bitcoin, etc.)

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=0.19.0
click>=8.0.0

# Notebook and visualization
jupyter>=1.0.0
plotly>=5.0.0
ipywidgets>=7.6.0

# Testing and quality
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional: Deep learning (uncomment if needed)
# torch>=1.12.0
# tensorflow>=2.9.0

# Optional: Additional ML libraries
# xgboost>=1.6.0
# lightgbm>=3.3.0
# optuna>=3.0.0

# Optional: Time series analysis
# statsmodels>=0.13.0